pyproject.toml
[project]
name = "notecard-extractor"
version = "0.1.0"
description = "A tool to extract and crop notecard images from PDF files"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pypdf",
    "pillow",
    "typer",
]

[project.scripts]
extract-notecards = "notecard_extractor.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
